+++ src/ExportGold.hh
@@ -25,16 +25,16 @@ public:
 
     Mesh* mesh;
 
-    std::vector<int> tris;         // zone index list for 3-sided zones
-    std::vector<int> quads;        // same, for 4-sided zones
-    std::vector<int> others;       // same, for n-sided zones, n > 4
-    std::vector<int> mapzs;        // map: zone -> first side
+    std::vector<long long> tris;         // zone index list for 3-sided zones
+    std::vector<long long> quads;        // same, for 4-sided zones
+    std::vector<long long> others;       // same, for n-sided zones, n > 4
+    std::vector<long long> mapzs;        // map: zone -> first side
 
     // parallel info, meaningful on PE 0 only:
-    std::vector<int> pentris;      // number of tris on each PE
-    std::vector<int> penquads;     // same, for quads
-    std::vector<int> penothers;    // same, for others
-    int gntris, gnquads, gnothers; // total number across all PEs
+    std::vector<long long> pentris;      // number of tris on each PE
+    std::vector<long long> penquads;     // same, for quads
+    std::vector<long long> penothers;    // same, for others
+    long long gntris, gnquads, gnothers; // total number across all PEs
                                    //     of tris/quads/others
 
     ExportGold(Mesh* m);
